tests: refactor tests
